Converted Core to a header-only library
