--- Analysis/bin/BuildFile.xml
@@ -1,16 +1,0 @@
-<bin name="ShuffleMergeSpectral" file="ShuffleMergeSpectral.cxx">
-  <use name="boost_program_options"/>
-  <use name="TauMLTools/Core"/>
-</bin>
-<bin name="CreateSpectralHists" file="CreateSpectralHists.cxx">
-  <use name="boost_program_options"/>
-  <use name="TauMLTools/Core"/>
-</bin>
-<bin name="MergeTuples" file="MergeTuples.cxx">
-    <use name="boost_program_options"/>
-    <use name="TauMLTools/Core"/>
-</bin>
-<bin name="SimpleMergeMix" file="SimpleMergeMix.cxx">
-    <use name="boost_program_options"/>
-    <use name="TauMLTools/Core"/>
-</bin>
