--- Core/BuildFile.xml
@@ -1,5 +1,0 @@
-<use name="boost"/>
-<use name="boost_filesystem"/>
-<use name="boost_regex"/>
-<use name="root"/>
-<export> <lib name="1"/> </export>
