+++ Core/interface/EnumNameMap.h
@@ -1,5 +1,5 @@
 /*! Definition of the I/O operators for enum <-> string conversion.
-This file is part of https://github.com/hh-italian-group/TauMLTools. */
+This file is part of https://github.com/cms-tau-pog/TauMLTools. */
 
 #pragma once
