+++ Core/interface/NumericPrimitives.h
@@ -1,5 +1,5 @@
 /*! Definition of the primitives that extend CERN ROOT functionality.
-This file is part of https://github.com/hh-italian-group/TauMLTools. */
+This file is part of https://github.com/cms-tau-pog/TauMLTools. */
 
 #pragma once
