+++ Core/interface/StatEstimators.h
@@ -1,5 +1,5 @@
 /*! Definition of data driven statistical estimators.
-This file is part of https://github.com/hh-italian-group/TauMLTools. */
+This file is part of https://github.com/cms-tau-pog/TauMLTools. */
 
 #pragma once
