+++ Core/interface/Tools.h
@@ -1,5 +1,5 @@
 /*! Common tools and definitions suitable for general purposes.
-This file is part of https://github.com/hh-italian-group/TauMLTools. */
+This file is part of https://github.com/cms-tau-pog/TauMLTools. */
 
 #pragma once
