+++ Core/interface/map_vec.h
@@ -1,5 +1,5 @@
 /*! Definition of the map class that keeps a trace of the insertion order.
-This file is part of https://github.com/hh-italian-group/TauMLTools. */
+This file is part of https://github.com/cms-tau-pog/TauMLTools. */
 
 #pragma once
